+++ semcore/card/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [5.43.1] - 2024-11-01
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/utils` [4.42.0 ~> 4.43.0], `@semcore/button` [5.39.0 ~> 5.39.1], `@semcore/icon` [4.51.0 ~> 4.52.0]).
+
 ## [5.43.0] - 2024-10-28
 
 ### Changed