+++ semcore/select/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.48.2] - 2024-11-01
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/base-trigger` [4.47.0 ~> 4.48.0], `@semcore/button` [5.39.0 ~> 5.39.1], `@semcore/checkbox` [7.40.0 ~> 7.41.0], `@semcore/divider` [4.36.0 ~> 4.37.0], `@semcore/dropdown-menu` [4.42.1 ~> 4.43.0], `@semcore/dropdown` [4.40.1 ~> 4.41.0], `@semcore/popper` [5.43.1 ~> 5.44.0], `@semcore/flex-box` [5.37.0 ~> 5.38.0], `@semcore/icon` [4.51.0 ~> 4.52.0], `@semcore/input` [4.38.0 ~> 4.39.0], `@semcore/utils` [4.42.0 ~> 4.43.0], `@semcore/core` [2.35.0 ~> 2.36.0]).
+
 ## [4.48.1] - 2024-10-28
 
 ### Changed