+++ semcore/tab-line/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.37.0] - 2024-11-01
+
+### Changed
+
+- Version minor update due to children dependencies update (`@semcore/flex-box` [5.37.0 ~> 5.38.0], `@semcore/neighbor-location` [4.35.0 ~> 4.36.0], `@semcore/utils` [4.42.0 ~> 4.43.0], `@semcore/core` [2.35.0 ~> 2.36.0]).
+
 ## [4.36.0] - 2024-10-28
 
 ### Changed